Extended Transaction (#36)
- Adds optional fields to TxIn 
  - `satoshis: Option<u64>`
  - `unlocking_script: Option<Script>`
- Adds convenience methods for querying Total Satoshis in and Total Satoshis out
  - `Tx.satoshisIn()` - Returns the total amount of satoshis in the Inputs. If any of the inputs have no specified satoshis, method returns null
  - `Tx.satoshisOut()`  - Returns the total amount of satoshis in the Outputs.
- Adds ability to match specific TxIns and TxOuts based on a "MatchCriteria"
  - `Tx.matchInput()` - Returns index of the first input that matches the given criteria
  - `Tx.matchOutput()` - Returns index of the first output that matches the given criteria
  - `Tx.matchInputs()` - Returns all indices of inputs that match the given criteria
  - `Tx.matchOutputs()` - Returns all indices of outputs that match the given criteria

Satoshis field on TxIn makes it easy to do calculations and signing for transactions that get passed around without having to consult a UTXO store. If you serialise the Tx to JSON or toCompactBytes() (flexbuffers), it adds the extra fields when available.

Unlocking Script field on TxIn allows you to pass around partially signed transactions without having to continually change the script sig. This field will automatically be prepended to the script_sig when calling .toBytes() or .toHex() to enable a convenient way to export the transaction to Bitcoin network format.

Both of these features are completely opt-in. Those who want a more vanilla Bitcoin experience, do not have to use them. This feature is completely backwards compatible.